put components in their own page
